requirements.txt
PILLOW
torchvision
torch
numpy